Refactor account settings page to show "Les Toevoegen" button for owners.
This commit modifies the account settings page to display a "Les Toevoegen" button for users with the role of "eigenaar". Previously, only users with the role of "instructeur" could see this button. The code changes were made in pages/account_settings.php. Additionally, include/meldingen.php was deleted as it was no longer needed.