geen idee wat ik hier heb gedaan
Added an if statement to check if $selected_date is empty after the submit button is pressed. If it's empty, display a message asking the user to select a date. Also added code to update the opmerking field in the les table when reden_afmelding equals 'anders'. Finally, closed the mysqli connection and added an event listener for date input changes.