style.css: Adjusted margin and width of tarieven_card
The margin of the tarieven_card has been adjusted to 1rem on top, 4rem on the right, 1rem on the bottom, and 1rem on the left. The width of left_side_tarieven_card has been changed from 24vw to min-width:16vw. Additionally, white-space property has been added to .tarieven_card class.