Refactor popup button styling and add close icon
This commit refactors the styling of the popup button to remove its gray color and make it transparent. It also adds a close icon to the button using CSS content property. The size of the icon is increased to 50px, and its color is set to match the secondary color of the website. Finally, we adjust its position by setting a relative position and top offset.