Refactor CSS style, add subscription functionality and redirect to account settings page.
Refactor CSS style, add subscription functionality and redirect to account settings page.

This commit refactors the CSS style by removing the label styling and adding it to a specific ID. It also adds subscription functionality to the website by allowing users to select a lespakket and update their aantallessen. Finally, it redirects users to the account settings page after logging in or registering, specifically to the Subscription section.